+++ lib/mobile/flutter_gemma_mobile_model_manager.dart
@@ -1,13 +1,13 @@
 part of 'flutter_gemma_mobile.dart';
 
-const _modelPath = 'model.bin';
-const _loraPath = 'lora.bin';
-
 class MobileModelManager extends ModelFileManager {
   MobileModelManager({
     required this.onDeleteModel,
     required this.onDeleteLora,
   });
+
+  var _loraPath = 'lora.bin';
+  var _modelPath = 'model.bin';
 
   final AsyncCallback onDeleteModel;
   final AsyncCallback onDeleteLora;
@@ -24,6 +24,7 @@ class MobileModelManager extends ModelFileManager {
   Future<File> get _modelFile async {
     if (_userSetModelPath case String path) return File(path);
     final directory = await _docsDirectory;
+
     return File('${directory.path}/$_modelPath');
   }
 
@@ -49,7 +50,8 @@ class MobileModelManager extends ModelFileManager {
         targetPath: target,
       );
 
-  Future<void> _loadAsset(String path, String target) => _largeFileHandler.copyAssetToLocalStorage(
+  Future<void> _loadAsset(String path, String target) =>
+      _largeFileHandler.copyAssetToLocalStorage(
         assetName: path,
         targetPath: target,
       );
@@ -69,7 +71,9 @@ class MobileModelManager extends ModelFileManager {
   Future<bool> get _shouldLoadModel async {
     final completer = _modelCompleter;
     if (completer != null) {
-      return completer.isCompleted ? false : throw Exception('Model is already loading');
+      return completer.isCompleted
+          ? false
+          : throw Exception('Model is already loading');
     }
     final modelFile = await _modelFile;
     // Do not reload model if it's already exists.
@@ -80,7 +84,9 @@ class MobileModelManager extends ModelFileManager {
   Future<bool> get _shouldLoadLora async {
     final completer = _loraCompleter;
     if (completer != null) {
-      return completer.isCompleted ? false : throw Exception('Lora weights are already loading');
+      return completer.isCompleted
+          ? false
+          : throw Exception('Lora weights are already loading');
     }
     final loraFile = await _loraFile;
     // Do not reload lora weights if it's already exists.
@@ -106,7 +112,8 @@ class MobileModelManager extends ModelFileManager {
     }
   }
 
-  Stream<int> _loadModelWithProgressIfNeeded(Stream<int> Function() loadCallback) async* {
+  Stream<int> _loadModelWithProgressIfNeeded(
+      Stream<int> Function() loadCallback) async* {
     if (await _shouldLoadModel) {
       final completer = _modelCompleter = Completer<bool>();
       try {
@@ -131,10 +138,17 @@ class MobileModelManager extends ModelFileManager {
 
   @override
   Future<void> setModelPath(String path, {String? loraPath}) async {
-    await Future.wait([
-      _loadModelIfNeeded(() async => _userSetModelPath = path),
-      if (loraPath != null) _loadLoraIfNeeded(() async => _userSetLoraPath = loraPath),
-    ]);
+    _modelPath = path;
+    if (loraPath != null) {
+      _loraPath = loraPath;
+    }
+    if (await isModelInstalled) {
+      await Future.wait([
+        _loadModelIfNeeded(() async => _userSetModelPath = path),
+        if (loraPath != null)
+          _loadLoraIfNeeded(() async => _userSetLoraPath = loraPath),
+      ]);
+    }
   }
 
   @override
@@ -146,7 +160,8 @@ class MobileModelManager extends ModelFileManager {
   Future<void> installModelFromAsset(String path, {String? loraPath}) async {
     await Future.wait([
       _loadModelIfNeeded(() => _loadAsset(path, _modelPath)),
-      if (loraPath != null) _loadLoraIfNeeded(() => _loadAsset(loraPath, _loraPath)),
+      if (loraPath != null)
+        _loadLoraIfNeeded(() => _loadAsset(loraPath, _loraPath)),
     ]);
   }
 
@@ -154,12 +169,14 @@ class MobileModelManager extends ModelFileManager {
   Future<void> downloadModelFromNetwork(String url, {String? loraUrl}) async {
     await Future.wait([
       _loadModelIfNeeded(() => _loadNetwork(url, _modelPath)),
-      if (loraUrl != null) _loadLoraIfNeeded(() => _loadAsset(loraUrl, _loraPath)),
+      if (loraUrl != null)
+        _loadLoraIfNeeded(() => _loadAsset(loraUrl, _loraPath)),
     ]);
   }
 
   @override
-  Stream<int> installModelFromAssetWithProgress(String path, {String? loraPath}) async* {
+  Stream<int> installModelFromAssetWithProgress(String path,
+      {String? loraPath}) async* {
     yield* _loadModelWithProgressIfNeeded(() => _streamAsset(path, _modelPath));
     if (loraPath != null) {
       await installLoraWeightsFromAsset(loraPath);
@@ -167,8 +184,10 @@ class MobileModelManager extends ModelFileManager {
   }
 
   @override
-  Stream<int> downloadModelFromNetworkWithProgress(String url, {String? loraUrl}) async* {
-    yield* _loadModelWithProgressIfNeeded(() => _streamNetwork(url, _modelPath));
+  Stream<int> downloadModelFromNetworkWithProgress(String url,
+      {String? loraUrl}) async* {
+    yield* _loadModelWithProgressIfNeeded(
+        () => _streamNetwork(url, _modelPath));
     if (loraUrl != null) {
       await downloadLoraWeightsFromNetwork(loraUrl);
     }