+++ lib/preferred_backend.dart
@@ -1,0 +1,5 @@
+enum PreferredBackend {
+  defaultBackend,
+  cpu,
+  gpu;
+}